# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "hiphouse API",
    "description": "크롤러가 채운 Postgres(artists/albums) CRUD 서버.",
    "version": "0.1.0"
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/albums/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "tags": ["albums"],
        "summary": "앨범 단건 조회",
        "responses": {
          "200": { "description": "앨범", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["albums"],
        "summary": "앨범 수정 (전체 교체)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } }
        },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["albums"],
        "summary": "앨범 삭제",
        "responses": {
          "204": { "description": "삭제됨" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": { "description": "잘못된 요청", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "없음", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Album": {
        "type": "object",
        "required": ["id", "name", "artists"],
        "properties": {
          "id": { "type": "string", "example": "4aawyAB9vmqN3uQ7FjRGTy" },
          "name": { "type": "string", "example": "Album Name" },
          "display_name": { "type": "string", "nullable": true, "example": "새삥", "description": "관리자가 등록한 한글 표시 이름. 없으면 name으로 표시하고, 검색 시 name과 함께 매칭됨" },
          "artists": {
            "type": "array",
            "description": "크레딧된 모든 아티스트 (Spotify 순서). album_artists 조인에서 집계.",
            "items": { "$ref": "#/components/schemas/Artist" }
          },
          "release_date": { "type": "string", "nullable": true, "example": "2024-05-01" },
          "year": { "type": "integer", "nullable": true, "example": 2024 },
          "album_type": { "type": "string", "nullable": true, "example": "album" },
          "total_tracks": { "type": "integer", "nullable": true, "example": 12 },
          "image_url": { "type": "string", "nullable": true },
          "spotify_url": { "type": "string", "nullable": true }
        }
      },
      "Artist": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string", "example": "1uNFoZAHBGtllmzznpCI3s" },
          "name": { "type": "string", "nullable": true, "example": "Artist Name" },
          "display_name": { "type": "string", "nullable": true, "example": "지코", "description": "관리자가 등록한 한글 표시 이름. 없으면 name으로 표시하고, 검색 시 name과 함께 매칭됨" },
          "image_url": { "type": "string", "nullable": true, "description": "Spotify 아티스트 사진 (없을 수 있음)" },
          "genres": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Spotify 아티스트 장르 (비어있을 수 있음)" },
          "spotify_url": { "type": "string", "nullable": true, "description": "Spotify 아티스트 프로필 URL" },
          "aliases": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "연관검색어 — 관리자가 등록한 검색 키워드 (예: 영문명의 한글 표기). 검색 시 name과 함께 매칭됨" },
          "followers": { "type": "integer", "nullable": true, "description": "Spotify 팔로워 수 — 크롤러/enrich가 채움 (아직 수집 전이면 null). 화면 표시용이 아니라 관리자 enrich 판단용" },
          "follower_count": { "type": "integer", "description": "서비스 내부 팔로워 수 — follows 테이블 집계" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  }
}